requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
openai>=1.54.0
pydantic>=2.6
PyYAML>=6.0
pyarrow>=15.0
pytest>=8.0
pillow>=10.0
